+++ packages/ui/src/core/primitives/text/styles.ts
@@ -64,6 +64,8 @@ export function textBaseStyle(
       /* Certain popular CSS libraries changes the defaults for SVG display */
       /* Make sure SVGs are rendered as inline elements */
       display: inline;
+      /* ...and aren't squeezed by resets that set max-width: 100% */
+      max-width: unset;
     }
 
     & [data-sanity-icon] {