+++ packages/themer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sanity/themer
 
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies [[`e93ebeb`](https://github.com/sanity-io/ui/commit/e93ebebc030d3f6e4b6872496f8ccc30247138f6)]:
+  - @sanity/ui@4.0.3
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [[`adf0894`](https://github.com/sanity-io/ui/commit/adf0894ba356d15b2af522f4f86fe532b09b6f21)]:
+  - @sanity/ui@4.0.2
+
 ## 0.3.3
 
 ### Patch Changes