+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sanity/ui
 
+## 4.0.3
+
+### Patch Changes
+
+- [#2702](https://github.com/sanity-io/ui/pull/2702) [`e93ebeb`](https://github.com/sanity-io/ui/commit/e93ebebc030d3f6e4b6872496f8ccc30247138f6) Thanks [@stipsan](https://github.com/stipsan)! - Unset `max-width` on inline SVGs in `Text`, `Label`, `Heading`, and `Code` so CSS resets that set `max-width: 100%` no longer squash icons.
+
+## 4.0.2
+
+### Patch Changes
+
+- [#2602](https://github.com/sanity-io/ui/pull/2602) [`adf0894`](https://github.com/sanity-io/ui/commit/adf0894ba356d15b2af522f4f86fe532b09b6f21) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency motion to ^13.1.0
+
 ## 4.0.1
 
 ### Patch Changes